requirements.txt
wxPython==4.2.4
opencv-python==4.13.0.90
PyAudio==0.2.14
numpy==1.23.5
sounddevice==0.5.5
soundfile==0.13.1
groq==1.0.0
pygame==2.6.1
SpeechRecognition==3.14.5
gTTS==2.5.4
torch==2.8.0
librosa==0.8.1
scipy==1.13.1
tqdm==4.67.1
deepface==0.0.97
openai-whisper==20250625
pyttsx3==2.99